requirements.txt
haystack-ai==2.4.0
fastapi==0.95.2
uvicorn[standard]==0.22.0